pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-qa-system"
version = "0.1.0"
description = "Python RAG QA System with ChromaDB and LlamaIndex"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LlamaIndex core and integrations
    "llama-index-core>=0.10.0",
    "llama-index-llms-dashscope>=0.1.0",
    "llama-index-embeddings-dashscope>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-index-readers-file>=0.1.0",

    # ChromaDB
    "chromadb>=0.4.0",

    # PDF parsing
    "docling>=2.0.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"
